move manipluation of list_display and list_filter
move the manipulation of list_display and list_filter to overridden get_list_*
methods. The ModelAdmin seems to get instantiated when executing management
commands and the user_count query throws an exception table doesn't exist if
the database hasn't been created. This means that it's not possible to actually
create the database.